* ftpcmd.y, ftpd.c: Cast argument to isspace()/isdigit() to int
        * secure.h (SESSION): Pass pointer to C_Block as expected by
        krb5_mk_priv().

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13351 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/appl/gssftp/ftpd/ftpd.c
@@ -882,10 +882,10 @@ checkuser(name)
 			     int i = strlen(name) + 1;
 			     
 			     /* Make sure foo doesn't match foobar */
-			     if (line[i] == '\0' || !isspace(line[i]))
+			     if (line[i] == '\0' || !isspace((int) line[i]))
 			          continue;
 			     /* Ignore whitespace */
-			     while (isspace(line[++i]));
+			     while (isspace((int) line[++i]));
 
 			     if (strcmp(&line[i], "restrict") == 0)
 			          return (-1);
